requirements.txt
# ----------------------------------------
# Python Version
# ----------------------------------------
python>=3.9

# ----------------------------------------
# Core Libraries
# ----------------------------------------
requests>=2.25.1
beautifulsoup4>=4.9.3
newspaper3k>=0.2.8
python-dotenv>=0.19.2
pyyaml>=6.0

# ----------------------------------------
# Web Scraping & Browser Automation
# ----------------------------------------
selenium>=4.1.0
webdriver-manager>=3.8.5

# ----------------------------------------
# Social Media & Content APIs
# ----------------------------------------
praw>=7.7.0              # Reddit API wrapper
snscrape>=0.6.0          # Twitter/X scraping
quora-scraper>=0.1.0     # Quora scraping

# ----------------------------------------
# NLP & AI
# ----------------------------------------
spacy>=3.5.0
transformers>=4.26.1
sentence-transformers>=2.2.0
nltk>=3.8.1
scikit-learn>=1.2.1
textblob>=0.17.1
openai>=0.27.0

# ----------------------------------------
# Data Processing
# ----------------------------------------
pandas>=1.3.5
numpy>=1.21.5
scipy>=1.7.3

# ----------------------------------------
# Output Generation & Presentation
# ----------------------------------------
markdown>=3.4.1
markdown2>=2.4.0
fpdf>=1.7.2
pdfkit>=1.0.0
jinja2>=3.1.2
rich>=12.0.0             # For pretty CLI output

# ----------------------------------------
# Web API Framework
# ----------------------------------------
fastapi>=0.92.0
uvicorn>=0.20.0

# ----------------------------------------
# Testing
# ----------------------------------------
pytest>=7.0.1
pytest-asyncio>=0.18.3
pytest-cov>=3.0.0
responses>=0.22.0
pytest-mock>=3.10.0

# ----------------------------------------
# Additional Dependencies
# ----------------------------------------
aiohttp>=3.8.1
feedparser>=6.0.8
redis>=4.3.4
loguru>=0.6.0
html5lib>=1.1
isodate==0.6.1

# Research and academic scrapers
arxiv>=1.4.2
scholarly>=1.7.11

# Database and caching
pymongo>=4.3.3

# Utilities
tqdm>=4.64.0
